alerta al borrar

+++ src/app/players/players.component.ts
@@ -77,10 +77,14 @@ export class PlayersComponent implements OnInit {
   //   this.playerService.resetDatabase(InMemoryDataService);
   // }
   onDeletePlayer(playerId: string) {
+    const confirmDelete = window.confirm('¿Estás seguro de que quieres eliminar este jugador?');
+
+    if (!confirmDelete) return;
+
     this.playerService.deletePlayer(playerId).subscribe({
       next: () => {
         console.log(`Jugador con ID ${playerId} eliminado correctamente.`);
-        this.filteredPlayers$(); // 🔁 Volvemos a cargar la lista tras el borrado
+        this.filteredPlayers$(); // 🔁 Recargar la lista tras el borrado
       },
       error: (err) => {
         console.error('Error al eliminar el jugador:', err);
